Added warning to ListLayout when options.spacing is non-numeric
The CollectionLayout supports array-based spacing options, and the ListLayout does not.

If an array spacing option is given to ListLayout (after switching from CollectionLayout back to ListLayout for example), it does not warn or throw an exception, but rather doesn't show any renderables without notice. This commit fixes that behaviour by letting the developer know that the spacing should be numeric.